Review pass: shared cell tags, float64 heads, device-safe spectral helpers, cumulative BN calibration, explicit reference budget and leader decay, tests
- bilevel_lt/tags.py: one torch-free tag builder used by the ladder driver and
  scripts/sweep_eta.py (imbalance and ref_steps now part of the tag)
- heads.init draws float64 regardless of the default dtype
- estimators: lam_max/scipy_operator/lanczos take dtype and device from the caller;
  the MINRES arm is named minres in both cert and loop
- resnet.calibrate_bn uses the cumulative average (momentum None) so the reset
  values carry no weight
- ladder: --ref_steps and --tau_grid; k_list default 200,600
- autobalance: --outer_wd exposed (default 1e-4 as published); docstring states that
  H and B are of the cross-entropy term only, that sigma(delta) starts at 0.5, and
  that minibatches are proper epochs
- tests: tag agreement between driver and sweep tool, hypergradient sign and scale
  against finite differences on a convex synthetic problem, SGD step against
  torch.optim.SGD, Flattener round trip

+++ bilevel_lt/heads.py
@@ -49,8 +49,8 @@ def build_head(kind: str, k: int, C: int):
         for i in range(n_layers):
             nout, nin = dims[i + 1], dims[i]
             bound = 1.0 / math.sqrt(nin)
-            parts.append((torch.rand(nout * nin, generator=g) * 2 - 1) * bound)
-            parts.append((torch.rand(nout, generator=g) * 2 - 1) * bound)
+            parts.append((torch.rand(nout * nin, generator=g, dtype=torch.float64) * 2 - 1) * bound)
+            parts.append((torch.rand(nout, generator=g, dtype=torch.float64) * 2 - 1) * bound)
         return torch.cat(parts)
 
     return forward, init, p, dims
